+++ plugins/command-manager/src/main/java/com/wazuh/commandmanager/utils/Search.java
@@ -20,12 +20,12 @@ import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 import org.opensearch.action.search.SearchRequest;
 import org.opensearch.action.search.SearchResponse;
-import org.opensearch.transport.client.support.AbstractClient;
 import org.opensearch.core.action.ActionListener;
 import org.opensearch.index.query.BoolQueryBuilder;
 import org.opensearch.index.query.QueryBuilders;
 import org.opensearch.search.SearchHits;
 import org.opensearch.search.builder.SearchSourceBuilder;
+import org.opensearch.transport.client.support.AbstractClient;
 
 import java.util.*;
 import java.util.concurrent.CountDownLatch;